Update Resources.cfg
Added cost to karbonite so it could be sold to get more funds. The prices is a bit higher than fuel and oxidizer so it doesn't exactly make sense from an economic point of view (unless we assume it can be used for more than just fuel conversion). Point is it has value now. Cost tweaking will probably need to happen as well.