„Remaining time“ addition for Karbonite Generator
In case Rover accepts my PR for the USI_Converter I thought it would be
nice to have a consistent experience throughout the USI range and added
a similar system to the generator.

Issue: once fully loaded and with only a tiny drain the guiName
flickers. May be better to remove the “curr. rate” “last rate” stuff
altogether and replace it with just “Remaining” like in the USI. Not
sure tbh, may Rover decide.

Rover: again, do whatever you want with this.